Move the LabelTest utility class to a utils.py file.
