+++ project/labels/tests/test_labelsets.py
@@ -11,7 +11,7 @@ from accounts.utils import get_robot_user
 from images.model_utils import PointGen
 from lib.tests.utils import BasePermissionTest, sample_image_as_file
 from ..models import Label
-from .test_labels import LabelTest
+from .utils import LabelTest
 
 
 class PermissionTest(BasePermissionTest):
